refactor(server): isolate account session lifecycle
